refactor: group use cases by processes

+++ app/user/application/user_creation_process/user_creation_use_case.py
@@ -9,44 +9,52 @@ from app.core.services.pin_service import generate_pin
 from app.core.services.email_service import send_email
 from datetime import datetime, timezone, timedelta
 
-
 class UserCreationUseCase:
     def __init__(self, db: Session):
         self.db = db
         self.user_repository = UserRepository(db)
 
-    def execute(self, user_data: UserCreate) -> str:
+    def execute(self, user_data: UserCreate) -> dict:
         # Verificar si el usuario ya existe
         existing_user = self.user_repository.get_user_by_email(user_data.email)
+        
         if existing_user:
+            # Verificar el estado del usuario
+            active_state_id = self.user_repository.get_active_user_state_id()
+            if existing_user.state_id == active_state_id:
+                raise DomainException(
+                    message="La cuenta ya está confirmada y activa.",
+                    status_code=status.HTTP_400_BAD_REQUEST
+                )
+                
+            # Verificar si el usuario ha sido eliminado
+            inactive_state_id = self.user_repository.get_inactive_user_state_id()
+            if existing_user.state_id == inactive_state_id:
+                raise DomainException(
+                    message="El usuario fue eliminado del sistema.",
+                    status_code=status.HTTP_410_GONE
+                )
+
+            # Verificar si la cuenta del usuario está bloqueada        
+            if existing_user.locked_until:
+                existing_user.locked_until = existing_user.locked_until.replace(tzinfo=timezone.utc)
+
+            locked_state_id = self.user_repository.get_locked_user_state_id()
+            if existing_user.state_id == locked_state_id and existing_user.locked_until > datetime.now(timezone.utc):
+                time_left = existing_user.locked_until - datetime.now(timezone.utc)
+                minutos_restantes = time_left.seconds // 60
+                raise DomainException(
+                    message=f"Su cuenta está bloqueada. Intente nuevamente en {minutos_restantes} minutos.",
+                    status_code=status.HTTP_403_FORBIDDEN
+                )
+            
             # Verificar si el usuario tiene una confirmación pendiente
             pending_confirmation = self.user_repository.get_user_pending_confirmation(existing_user.id)
             if pending_confirmation:
-                # Eliminar usuario y confirmación pendiente
+                # Eliminar usuario y confirmación pendiente para crearlos de nuevo
                 self.user_repository.delete_user_confirmations(existing_user.id)
                 self.user_repository.delete_user(existing_user)
-                print("Usuario existente eliminado para permitir la re-creación.")
-            else:
-                raise DomainException(
-                    message="El correo electrónico proporcionado ya está en uso.",
-                    status_code=status.HTTP_409_CONFLICT,
-                )
 
-        # Crear nuevo usuario
-        new_user = self.create_user(user_data)
-
-        # Intentar crear la confirmación y enviar el correo
-        if not self.create_and_send_confirmation(new_user):
-            # Si falla la confirmación, eliminar el usuario
-            self.user_repository.delete_user(new_user)
-            raise DomainException(
-                message="Error al crear la confirmación de usuario.",
-                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            )
-
-        return "Usuario creado. Por favor, revisa tu email para confirmar el registro."
-
-    def create_user(self, user_data: UserCreate) -> User:
         # Hash del password
         hashed_password = hash_password(user_data.password)
 
@@ -72,17 +80,28 @@ class UserCreationUseCase:
         unconfirmed_role = self.user_repository.get_unconfirmed_user_role()
         if unconfirmed_role:
             if not self.user_repository.assign_role_to_user(created_user.id, unconfirmed_role.id):
+                self.user_repository.delete_user(created_user)  # Eliminar el usuario si falla la asignación del rol
                 raise DomainException(
                     message="No se pudo asignar el rol de Usuario No Confirmado.",
                     status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
                 )
         else:
+            self.user_repository.delete_user(created_user)  # Eliminar el usuario si no se encuentra el rol
             raise DomainException(
                 message="No se pudo encontrar el rol de Usuario No Confirmado.",
                 status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
             )
 
-        return created_user
+        # Intentar crear la confirmación y enviar el correo
+        if not self.create_and_send_confirmation(created_user):
+            # Si falla la confirmación, eliminar el usuario
+            self.user_repository.delete_user(created_user)
+            raise DomainException(
+                message="Error al crear la confirmación de usuario.",
+                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
+            )
+
+        return {"message":"Usuario creado. Por favor, revisa tu email para confirmar el registro."}
 
     def create_and_send_confirmation(self, user: User) -> bool:
         try:
